Добавлена валидация
Исправлены ошибки, при которых, программа зацикливалась из-за некорректного ввода

+++ laba_1/custom_math.c
@@ -1,7 +1,5 @@
-#include "custom_math.h"
-
-long double power(const double x,const int power){
-    double result = 1.0;
+long double power(const double x, const int power){
+    long double result = 1.0;
     for (int i = 0; i < power; i++) {
         result *= x;
     }
